+++ src/desk/desk-helpers.js
@@ -1,9 +1,5 @@
 import { CHARACTERISTICS } from './desk-constants';
 
-export const sleep = (ms) => {
-  return new Promise(resolve => setTimeout(resolve, ms));
-};
-
 class DeskHelpers {
   getHeightCharacteristic = (characteristics) => {
     return this.getCharacteristicByUUID(characteristics, CHARACTERISTICS.height.uuidNoDashesLowCase);
