+++ src/desk/desk-manager.js
