+++ src/shared/helpers.js
@@ -1,0 +1,4 @@
+
+export const sleep = (ms) => {
+  return new Promise(resolve => setTimeout(resolve, ms));
+};
